+++ lib/src/analytics/analytics.dart
@@ -1,7 +1,5 @@
 import 'package:loglytics/loglytics.dart';
 
-import '../../loglytics.dart';
-
 /// Base class to be inherited when specifying analytics for a specific feature or part of your project.
 ///
 /// Comes with an [AnalyticsService] to facilitate easy logging of analytics.