+++ lib/src/loglytics/event_bus.dart
@@ -3,7 +3,7 @@ part of 'loglytics.dart';
 /// Used to make sure that events that are fired will be sent in chronological order.
 class EventBus {
   EventBus._();
-  static late final EventBus _eventBus = EventBus._();
+  static final EventBus _eventBus = EventBus._();
   factory EventBus() => _eventBus;
 
   late final StreamController<Future<void>> _crashReports =